quotation_microservice/requirements.txt
fastapi~=0.111.0
uvicorn[standard]~=0.30.1
openai~=1.35.13
pydantic-settings~=2.3.4
python-dotenv~=1.0.1